Move iterator utils to dedicated module (in preparation for other PR with more utils)
