+++ src/gt4py/next/iterator/ir_utils/common_pattern_matcher.py
